+++ crack-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "crack-core"
-version = "0.3.5-rc.0"
+version = "0.3.5"
 authors = ["Cycle Five <cycle.five@proton.me>"]
 edition = "2021"
 description = "Core module for the cracking smoking, discord-music-bot Cracktunes."
@@ -81,7 +81,7 @@ mockall = "0.12.1"
 tungstenite = "0.21.0"
 async-tungstenite = "0.25.1"
 sqlx = { version = "0.7.4", features = [
-    "runtime-tokio",
+    "runtime-tokio-rustls",
     "tls-rustls",
     "macros",
     "postgres",
