+++ crack-core/src/commands/music/grab.rs
@@ -1,15 +1,7 @@
-use poise::serenity_prelude::{ChannelId, Message};
-use serenity::{builder::CreateMessage, http::Http};
-use songbird::input::AuxMetadata;
-use songbird::Call;
-use std::sync::Arc;
-use std::time::Duration;
-use tokio::sync::Mutex;
+use crate::utils::send_now_playing;
+use crate::{errors::CrackedError, Context, Error};
 
-use crate::utils::create_now_playing_embed_metadata;
-use crate::utils::get_requesting_user;
-use crate::{errors::CrackedError, interface::create_now_playing_embed, Context, Error};
-
+/// interface::create_now_playing_embed,
 /// Send the current tack to your DMs.
 #[cfg(not(tarpaulin_include))]
 #[poise::command(slash_command, prefix_command, aliases("save"), guild_only)]
@@ -40,45 +32,4 @@ pub async fn grab(ctx: Context<'_>) -> Result<(), Error> {
     ctx.data().add_msg_to_cache(guild_id, msg);
 
     Ok(())
-}
-
-/// Send the current track information as an ebmed to the given channel.
-#[cfg(not(tarpaulin_include))]
-pub async fn send_now_playing(
-    channel: ChannelId,
-    http: Arc<Http>,
-    call: Arc<Mutex<Call>>,
-    cur_position: Option<Duration>,
-    metadata: Option<AuxMetadata>,
-) -> Result<Message, Error> {
-    tracing::warn!("locking mutex");
-    let mutex_guard = call.lock().await;
-    tracing::warn!("mutex locked");
-    let msg: CreateMessage = match mutex_guard.queue().current() {
-        Some(track_handle) => {
-            tracing::warn!("track handle found, dropping mutex guard");
-            drop(mutex_guard);
-            let requesting_user = get_requesting_user(&track_handle).await;
-            let embed = if let Some(metadata2) = metadata {
-                create_now_playing_embed_metadata(
-                    requesting_user.ok(),
-                    cur_position,
-                    crate::commands::MyAuxMetadata::Data(metadata2),
-                )
-            } else {
-                create_now_playing_embed(&track_handle).await
-            };
-            CreateMessage::new().embed(embed)
-        },
-        None => {
-            tracing::warn!("track handle not found, dropping mutex guard");
-            drop(mutex_guard);
-            CreateMessage::new().content("Nothing playing")
-        },
-    };
-    tracing::warn!("sending message: {:?}", msg);
-    channel
-        .send_message(Arc::clone(&http), msg)
-        .await
-        .map_err(|e| e.into())
 }
