+++ cracktunes/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Cycle Five <cycle.five@proton.me>"]
 name = "cracktunes"
-version = "0.3.5-rc.0"
+version = "0.3.5"
 description = "Cracktunes is a hassle-free, highly performant, host-it-yourself, cracking smoking, discord-music-bot."
 publish = true
 edition = "2021"
@@ -45,11 +45,11 @@ sqlx = { workspace = true, optional = true }
 
 async-tungstenite = "0.25"
 tungstenite = "0.21"
-async-trait = { version = "0.1" }
-mockall = { version = "0.12" }
-config-file = { version = "0.2", features = ["json"] }
-dotenv = "0.15"
-colored = "2.1"
+async-trait = { version = "0.1.80" }
+mockall = { version = "0.12.1" }
+config-file = { version = "0.2.3", features = ["json"] }
+dotenv = "0.15.0"
+colored = "2.1.0"
 # Figure this one out better
 
 # crack-telemetry
@@ -60,10 +60,15 @@ opentelemetry-otlp = { version = "0.15", features = [
     "trace",
     #    "grpc-tonic",
 ], optional = true }
-opentelemetry_sdk = { version = "0.22", features = [
+#
+# which one of these is best?
+# rt-tokio: Spawn telemetry tasks using tokio’s multi-thread runtime.
+# rt-tokio-current-thread: Spawn telemetry tasks on a separate runtime so that the main runtime won’t be blocked.
+#
+opentelemetry_sdk = { version = "0.22.1", features = [
     "rt-tokio",
 ], optional = true }
-opentelemetry-semantic-conventions = { version = "0.14", optional = true }
+opentelemetry-semantic-conventions = { version = "0.14.0", optional = true }
 # tonic = { version = "0.11", optional = true }
 tracing = { workspace = true }
 tracing-opentelemetry = { version = "0.23.0", optional = true }
@@ -75,8 +80,8 @@ tracing-subscriber = { version = "0.3.18", features = [
 tracing-futures = { version = "0.2.5", optional = true }
 
 # crack-metrics
-prometheus = { version = "0.13", features = ["process"], optional = true }
-warp = { version = "0.3", features = ["tls"], optional = true }
+prometheus = { version = "0.13.3", features = ["process"], optional = true }
+warp = { version = "0.3.7", features = ["tls"], optional = true }
 
 # # Is this even needed? What did I add it for?
 # This is needed for the tests to run in the IDE
